fixed right margin in the footer
Somehow, the `rightskip=\beamer@rightmargin` does not work any more for the footer. Without this fix, there is no space between the right edge of the frame and the frame number in the bottom right corner. I created a workaround by making the beamercolorbox smaller instead of using the rightskip.